[Billy] refactor directory structure
