+++ src/main/java/com/oocl/easyparkbackend/Employee/Exception/ClerkEmailAndPhoneNumberNotNullException.java
@@ -1,4 +1,4 @@
-package com.oocl.easyparkbackend.Clerk.Exception;
+package com.oocl.easyparkbackend.Employee.Exception;
 
 public class ClerkEmailAndPhoneNumberNotNullException extends RuntimeException {
