+++ src/main/java/com/oocl/easyparkbackend/Employee/Exception/NoSuchPositionException.java
@@ -1,4 +1,4 @@
-package com.oocl.easyparkbackend.Clerk.Exception;
+package com.oocl.easyparkbackend.Employee.Exception;
 
 public class NoSuchPositionException extends RuntimeException {
     @Override
