+++ src/test/java/com/oocl/easyparkbackend/Clerk/Repository/ClerkRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.oocl.easyparkbackend.Clerk.Repository;
 
-import com.oocl.easyparkbackend.Clerk.Entity.Clerk;
+import com.oocl.easyparkbackend.Employee.Entity.Clerk;
 import com.oocl.easyparkbackend.Manage.Entity.Manage;
 import com.oocl.easyparkbackend.Manage.Repository.ManageRepository;
 import com.oocl.easyparkbackend.common.vo.ClerkPosition;
